warnings have been punished

+++ src/lidar/lidarAnalize.cpp
@@ -182,8 +182,10 @@ position_float_t sol_eq_2cercle(position_float_t A, double RA, position_float_t 
     double y2 = A.y + a * (B.y - A.y) / d;
 
     // Points d'intersection
-    position_float_t P1 = {x2 + h * (B.y - A.y) / d, y2 - h * (B.x - A.x) / d};
-    position_float_t P2 = {x2 - h * (B.y - A.y) / d, y2 + h * (B.x - A.x) / d};
+    position_float_t P1 = {x2 + (float)(h * (B.y - A.y) / d), 
+                            y2 - (float)(h * (B.x - A.x) / d)};
+    position_float_t P2 = {x2 - (float)(h * (B.y - A.y) / d), 
+                            y2 + (float)(h * (B.x - A.x) / d)};
 
     if (in_table(P2)) return P2;
     if (in_table(P1)) return P1;
@@ -206,14 +208,18 @@ position_float_t position_estime(double alpha12, double alpha23, double alpha31,
     position_float_t M = {(bal_1.x + bal_2.x) / 2, (bal_1.y + bal_2.y) / 2}; // Calcul du milieu du segment AB
     double d = sqrt((bal_2.x - bal_1.x) * (bal_2.x - bal_1.x) + (bal_2.y - bal_1.y) * (bal_2.y - bal_1.y)); // Calcul de la distance entre A et B
     double h = sqrt(R1 * R1 - (d / 2) * (d / 2)); // Calcul de la distance du milieu au centre
-    position_float_t C = {M.x - h * (bal_2.y - bal_1.y) / d, M.y + h * (bal_2.x - bal_1.x) / d}; // Calcul des centres possibles
-    position_float_t C11 = {M.x + h * (bal_2.y - bal_1.y) / d, M.y - h * (bal_2.x - bal_1.x) / d}; // Calcul des centres possibles
+    position_float_t C = {M.x - (float)(h * (bal_2.y - bal_1.y) / d), 
+                            M.y + (float)(h * (bal_2.x - bal_1.x) / d)}; // Calcul des centres possibles
+    position_float_t C11 = {M.x + (float)(h * (bal_2.y - bal_1.y) / d), 
+                            M.y - (float)(h * (bal_2.x - bal_1.x) / d)}; // Calcul des centres possibles
 
     position_float_t M2 = {(bal_3.x + bal_2.x) / 2, (bal_3.y + bal_2.y) / 2}; // Calcul du milieu du segment AB
     double d2 = sqrt((bal_2.x - bal_3.x) * (bal_2.x - bal_3.x) + (bal_2.y - bal_3.y) * (bal_2.y - bal_3.y)); // Calcul de la distance entre A et B
     double h2 = sqrt(R2 * R2 - (d2 / 2) * (d2 / 2)); // Calcul de la distance du milieu au centre
-    position_float_t C2 = {M2.x - h2 * (bal_2.y - bal_3.y) / d2, M2.y + h2* (bal_2.x - bal_3.x) / d2}; // Calcul des centres possibles
-    position_float_t C22 = {M2.x + h2 * (bal_2.y - bal_3.y) / d2, M2.y - h2* (bal_2.x - bal_3.x) / d2}; // Calcul des centres possibles
+    position_float_t C2 = {M2.x - (float)(h2 * (bal_2.y - bal_3.y) / d2), 
+                            M2.y + (float)(h2* (bal_2.x - bal_3.x) / d2)}; // Calcul des centres possibles
+    position_float_t C22 = {M2.x + (float)(h2 * (bal_2.y - bal_3.y) / d2), 
+                            M2.y - (float)(h2* (bal_2.x - bal_3.x) / d2)}; // Calcul des centres possibles
 
     position_float_t robot[4];
     robot[0] = sol_eq_2cercle(C,R1,C2,R2);
